% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signatures.R
\name{check_and_edit_names_in_signatures_df}
\alias{check_and_edit_names_in_signatures_df}
\title{Check for missing name, NAs in name, or empty name in a data frame of signatures}
\usage{
check_and_edit_names_in_signatures_df(signatures_df)
}
\arguments{
\item{signatures_df}{}
}
\value{
signatures_df with compliant names
}
\description{
Check for missing name, NAs in name, or empty name in a data frame of signatures
}
\examples{
check_and_edit_names_in_signatures_df(signatures_df)
}
