% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare.R
\name{check_compare_df_sample_col_and_move_to_rowname}
\alias{check_compare_df_sample_col_and_move_to_rowname}
\title{Internal function to check if the sample name is a column or a rowname and move it to a rowname if it is a column.}
\usage{
check_compare_df_sample_col_and_move_to_rowname(compare_df)
}
\arguments{
\item{compare_df}{}
}
\value{
A tibble.
}
\description{
Internal function to check if the sample name is a column or a rowname and move it to a rowname if it is a column.
}
\examples{
check_compare_df_sample_col_and_move_to_rowname(compare_df)
}
