% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sourmash_outputs.R
\name{check_file_and_url_one}
\alias{check_file_and_url_one}
\title{Check if local file path or URL are valid}
\usage{
check_file_and_url_one(file)
}
\arguments{
\item{file}{File path or URL}
}
\description{
Check if local file path or URL are valid
}
