% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signatures.R
\name{check_uniform_parameters_in_signatures_df}
\alias{check_uniform_parameters_in_signatures_df}
\title{Check uniformity in sketch parameters}
\usage{
check_uniform_parameters_in_signatures_df(signatures_df)
}
\arguments{
\item{signatures}{A data.frame representing many signatures.}
}
\value{
NULL. Stops execution if parameters are not uniform.
}
\description{
Check uniformity in sketch parameters
}
\examples{
check_uniform_parameters_in_signatures_df(signatures)
}
