% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gather.R
\name{from_gather_to_upset_df}
\alias{from_gather_to_upset_df}
\title{Convert a data frame containing sourmash gather results into a upset plot compatible data frame}
\usage{
from_gather_to_upset_df(gather_df)
}
\arguments{
\item{gather_df}{A data frame of multiple sourmash gather results created by `read_gather()`.}
}
\value{
An upset plot compliant data frame
}
\description{
`from_gather_to_upset_df()` converts a data frame like that produced by `read_gather()` into a named list of vectors.
Each vector in the list is named by the query_name column in the input data frame and contains the genome accessions found in that sample.
The `query_name` column is used for sample names.
The returned data frame will have the genome accessions as row names and the query_names column names and can be plotted with `UpSetR::upset()` or `ComplexUpset::upset()`.
If plotting with Complex Upset, additional metadata can be added to the data frame (and therefore the plot) by joining on the values of the genome_accession rownames.
}
\examples{
\dontrun{
from_gather_to_upset_df(gather_df)
}
}
