% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signatures.R
\name{from_list_to_upset_df}
\alias{from_list_to_upset_df}
\title{Convert a list of vectors into an upset plot-compatible data.frame while conserving value names as row.names}
\usage{
from_list_to_upset_df(upset_list)
}
\arguments{
\item{upset_list}{A named list of vectors. Each name should be one set and each vector to should contain the values to be intersected.}
}
\value{
A data.frame.
}
\description{
`from_list_to_upset_df()` copies the functionality of UpSetR::fromList() and converts a list of named vectors to an upset plot compatible data frame.
It modifies the original function to return a data frame with row names that record the identity of the items in each vector.
This allows additional metadata to be joined to the resultant data frame using the values of the items in the list.
}
\examples{
sample1 <- c("a", "b", "c")
sample2 <- c("b", "c", "d")
upset_list <- list(sample1 = sample1, sample2 = sample2)
from_list_to_upset_df(upset_list)
}
