% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.R
\name{from_taxonomy_annotate_to_phyloseq}
\alias{from_taxonomy_annotate_to_phyloseq}
\title{Transform the output of sourmash taxonomy annotate into a phyloseq object}
\usage{
from_taxonomy_annotate_to_phyloseq(taxonomy_annotate_df, metadata_df = NULL)
}
\arguments{
\item{taxonomy_annotate_df}{Data frame containing outputs from sourmash taxonomy annotate.}

\item{metadata_df}{Optional. A data frame of metadata. The row names must match the values of `query_names` in the `taxonomy_annotate_df`. Any additional columns may be added.}
}
\value{
A phyloseq object.
}
\description{
`from_taxonomy_annotate_to_phyloseq()` transforms a data frame that contains sourmash taxonomy annotate results into a phyloseq object.
Counts are derived from the `n_unique_kmers` column, the abundance-weighted number of unique k-mers overlapping between a query and its match in the database.
}
\examples{
\dontrun{
from_taxonomy_annotate_to_phyloseq(taxonomy_annotate_df)
}
}
