% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.R
\name{from_taxonomy_annotate_to_tax_table}
\alias{from_taxonomy_annotate_to_tax_table}
\title{Transform the output of sourmash taxonomy annotate into a phyloseq tax table}
\usage{
from_taxonomy_annotate_to_tax_table(taxonomy_annotate_df)
}
\arguments{
\item{taxonomy_annotate_df}{Data frame containing outputs from sourmash taxonomy annotate.}
}
\value{
A phyloseq tax_table object.
}
\description{
Transform the output of sourmash taxonomy annotate into a phyloseq tax table
}
\examples{
\dontrun{
from_taxonomy_annotate_to_tax_table(taxonomy_annotate_df)
}
}
