% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxonomy_annotate.R
\name{from_taxonomy_annotate_to_upset_inputs}
\alias{from_taxonomy_annotate_to_upset_inputs}
\title{Transform a taxonomy annotate data frame into an upset plot compliant data frame}
\usage{
from_taxonomy_annotate_to_upset_inputs(
  taxonomy_annotate_df,
  tax_glom_level = NULL
)
}
\arguments{
\item{taxonomy_annotate_df}{Data frame containing outputs from sourmash taxonomy annotate. Can contain results from one or many runs of sourmash taxonomy annotate. Agglomeration occurs within each query.}

\item{tax_glom_level}{Optional character string specifying the taxonomic rank to agglomerate k-mer counts. Must be one of "domain", "phylum", "class", "order", "family", "genus", "species."}
}
\value{
A list. The first object in the list is an upset compliant data frame. The second is the taxonomy_annotate_df used to build the upset data frame. The last is the tax_glom_level.
}
\description{
`from_taxonomy_annotate_to_upset_inputs()` transforms a data frame with produced using read_taxonomy_annotate on many results produced by sourmash taxonomy annotate into a upset-compliant data frame.
The function can optionally agglomerate to different levels of taxonomic rank (e.g. phylum) and then produce the upset compliant data frame after agglomeration.
}
\examples{
\dontrun{
from_taxonomy_annotate_to_upset_inputs()
}
}
