% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset_utils.R
\name{from_upset_df_to_intersection_summary}
\alias{from_upset_df_to_intersection_summary}
\title{Summarize the size of intersections between samples in an upset compliant data frame}
\usage{
from_upset_df_to_intersection_summary(upset_df)
}
\arguments{
\item{upset_df}{An upset compliant data frame.}
}
\value{
A data frame.
}
\description{
`from_upset_df_to_intersection_summary()` reports the same information as upset plots but in a data frame.
It reports the intersection identity and the number of elements shared by the intersections.
}
\examples{
sample1 <- c("a", "b", "c")
sample2 <- c("b", "c", "d")
upset_list <- list(sample1 = sample1, sample2 = sample2)
upset_df <- from_list_to_upset_df(upset_list)
from_upset_df_to_intersection_summary(upset_df)
}
