% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sourmash_outputs.R
\name{get_scaled_for_max_hash}
\alias{get_scaled_for_max_hash}
\title{Calculate the scaled value for a signature from the max_hash value}
\usage{
get_scaled_for_max_hash(sig_max_hash)
}
\arguments{
\item{sig_max_hash}{Max hash value in a signature.}
}
\value{
Integer; a scaled value
}
\description{
Calculate the scaled value for a signature from the max_hash value
}
