% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gut_compare_df}
\alias{gut_compare_df}
\title{Gut microbiome sourmash compare data frame}
\format{
## `gut_compare_df`
A data frame with 6 rows and 7 columns:
\describe{
  \item{sample}{Sample name}
  \item{SRR5935765, SRR5936131, SRR5936197, SRR5946920, SRR5946923, SRR5947006}{Sequence Read Archive run accessions. Aligned values specify similarity between two samples.}
  ...
}
}
\source{
<https://github.com/Arcadia-Science/sourmashconsumr/blob/main/data-raw/00_sourmash_commands.sh>
}
\usage{
gut_compare_df
}
\description{
A similarlity matrix between six stool microbiome shotgun metagenome samples.
}
\keyword{datasets}
