% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gut_gather_df}
\alias{gut_gather_df}
\title{Gut microbiome sourmash gather data frame}
\format{
## `gut_gather_df`
A data frame with 1,062 rows and 30 columns:
\describe{
  \item{intersect_bp}{Numeric. Estimated number of intersected base pairs between a metagenome and a genome in a database.}
  \item{f_orig_query}{Numeric. Fraction of the original query that belongs to the match.}
  \item{f_match}{Numeric. Fraction of the matched genome in the leftover query.}
  \item{f_unique_to_query}{Numeric. Fraction of the query that uniquely belongs to the match.}
  \item{f_unique_weighted}{Numeric. Abundance-weighted fraction of the query that uniquely belongs to the match.}
  \item{average_abund}{Numeric. Average abundance of k-mers in the metagenome that were in the match.}
  \item{median_abund}{Numeric. Median abundance of k-mers in the metagenome that were in the match.}
  \item{std_abund}{Numeric. Standard deviation of abundance of k-mers in the metagenome that were in the match.}
  \item{filename}{Character. File path for the database on the computer that sourmash gather was executed on.}
  \item{name}{Character. Name of matched genome in the sourmash gather database.}
  \item{genome_accession}{Character. Genome accession solved by cutting of the name variable at the first space.}
  \item{md5}{Character. MD5 hash for the matched genome sketch.}
  \item{f_match_orig}{Numeric. Fraction of the matched genome in the original query prior to gather subtraction.}
  \item{unique_intersect_bp}{Numeric. Estimated number of uniquely intersected base pairs between a metagenome and a genome in a database.}
  \item{gather_result_rank}{Numeric. Rank of match in gather results.}
  \item{remaining_bp}{Numeric. Remaining base pairs in the query after the match is removed.}
  \item{query_filename}{Character. File name for the query derived from the query sketch.}
  \item{query_name}{Character. Name of the query.}
  \item{query_md5}{Character. MD5 hash for the query.}
  \item{query_bp}{Character. Number of base pairs in the query.}
  \item{ksize}{Character. K-mer size used for sourmash gather.}
  \item{moltype}{Character. Molecule type used for sourmash gather.}
  \item{scaled}{Numeric. Scaled value that sourmash gather was performed at.}
  \item{query_n_hashes}{Numeric. Number of hashes (k-mers) in the query.}
  \item{query_abundance}{Logical. Whether hash (k-mer) abundance information was a part of the query sketch.}
  \item{query_containment_ani}{Numeric. Containment between the query and the match.}
  \item{match_containment_ani}{Numeric. Containment between the match and the query.}
  \item{average_containment_ani}{Numeric. Average of the two containment metrics query_containment_ani and match_containment_ani.}
  \item{max_containment_ani}{Numeric. Maximum containment between the two containment metrics query_containment_ani and match_containment_ani.}
  \item{potential_false_negative}{Logical. Whether the containment estimate is a potential false negative.}
  ...
}
}
\source{
<https://github.com/Arcadia-Science/sourmashconsumr/blob/main/data-raw/00_sourmash_commands.sh>
}
\usage{
gut_gather_df
}
\description{
The results from running sourmash gather on each of six stool microbiome
metagenomes against the GTDB rs207 representatives database.
}
\keyword{datasets}
