% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gut_signatures_df}
\alias{gut_signatures_df}
\title{Gut microbiome sourmash signatures data frame}
\format{
## `gut_signatures_df`
A data frame with 15,339 rows and 16 columns:
\describe{
  \item{class}{Character. Class of the sketch.}
  \item{email}{Character. Email recorded in the sketch.}
  \item{hash_function}{Character. Hash function used for the sketch.}
  \item{filename}{Character. Name of the file that was sketched.}
  \item{name}{Character. Name of the sketch.}
  \item{license}{Character. License for the sketch.}
  \item{num}{Integer. Number of hashes specified to be in the sketch.}
  \item{ksize}{Integer. K-mer size represented in the sketch.}
  \item{seed}{Integer. Seed for the hash function.}
  \item{max_hash}{Numeric. Maximum hash.}
  \item{scaled}{Numeric. Scaled value for the sketch.}
  \item{mins}{Numeric. MinHash, or hash, that represents each k-mer sketched from the sample.}
  \item{md5sum}{Character. MD5 hash for the sketch.}
  \item{abundances}{Integer. Abundance of hash in sample.}
  \item{molecule}{Character. Molecule type represented by sketch.}
  \item{version}{Numeric. Sketch version.}
  ...
}
}
\source{
<https://github.com/Arcadia-Science/sourmashconsumr/blob/main/data-raw/00_sourmash_commands.sh>
}
\usage{
gut_signatures_df
}
\description{
Sourmash sketches from six stool microbiome shotgun metagenome samples parsed
into a data frame.
}
\keyword{datasets}
