% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxonomy_annotate.R
\name{make_agglom_cols}
\alias{make_agglom_cols}
\title{Helper function to define column names for taxonomy agglomeration}
\usage{
make_agglom_cols(tax_glom_level, with_query_name = FALSE)
}
\arguments{
\item{tax_glom_level}{Character. Valid options are "domain", "phylum", "class", "order", "family", "genus", and "species".}

\item{with_query_name}{Boolean indicating whether the column name "query_name" should be included.}
}
\value{
A character vector of column names to be used for agglomeration
}
\description{
Helper function to define column names for taxonomy agglomeration
}
\examples{
make_agglom_cols("species", with_query_name = FALSE)
}
