% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare.R
\name{make_compare_mds}
\alias{make_compare_mds}
\title{Perform an MDS analysis from a sourmash compare output and produce a tidy data.frame}
\usage{
make_compare_mds(compare_df)
}
\arguments{
\item{compare_df}{A tibble or data.frame produced by read_compare_csv() representing a (dis)similarity matrix output by sourmash compare.}
}
\value{
A data.frame.
}
\description{
Perform an MDS analysis from a sourmash compare output and produce a tidy data.frame
}
\examples{
make_compare_mds(compare_df)
}
