% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxonomy_annotate.R
\name{make_expression}
\alias{make_expression}
\title{Make an expression vector for italicizing a subset of a vector}
\usage{
make_expression(x, plain = NULL)
}
\arguments{
\item{x}{A character vector.}

\item{plain}{A character vector specifying which string to keep in regular (not italic) format.}
}
\value{
An expression-formatted vector.
}
\description{
Make an expression vector for italicizing a subset of a vector
}
