% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder.R
\name{pivot_wider_taxonomy_annotate}
\alias{pivot_wider_taxonomy_annotate}
\title{Title}
\usage{
pivot_wider_taxonomy_annotate(taxonomy_annotate_tibble)
}
\arguments{
\item{taxonomy_annotate_tibble}{}
}
\value{
A tibble in wide format. Variables expanded is n_unique_kmers and colnames are query_name.
}
\description{
Title
}
\examples{
pivot_wider_taxonomy_annoate(taxonomy_annotate_tibble)
}
