% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare.R
\name{plot_compare_heatmap}
\alias{plot_compare_heatmap}
\title{Plot a heatmap from a tibble of sourmash compare results.}
\usage{
plot_compare_heatmap(compare_df, seed = 42, ...)
}
\arguments{
\item{compare_df}{A tibble or data.frame produced by read_compare_csv() representing a (dis)similarity matrix output by sourmash compare.}

\item{seed}{Random seed for function. Default is 42. Ensures repeatable plot outputs.}

\item{...}{Arguments passed to base::heatmap().}
}
\value{
A plot of sourmash compare results.
}
\description{
Plot a heatmap from a tibble of sourmash compare results.
}
\examples{
\dontrun{
plot_compare_heatmap()
}
}
