% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare.R
\name{plot_compare_mds}
\alias{plot_compare_mds}
\title{Plot an MDS data.frame produced from the output of sourmash compare}
\usage{
plot_compare_mds(compare_mds, label = TRUE)
}
\arguments{
\item{compare_mds}{A data.frame produced using make_compare_mds()}

\item{label}{Boolean controlling whether sample labels are added to the plot. The default, TRUE, plots labels.}
}
\value{
A ggplot2 plot.
}
\description{
Plot an MDS data.frame produced from the output of sourmash compare
}
\examples{
\dontrun{
plot_compare_mds(compare_mds)
}
}
