% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gather.R
\name{plot_gather_classified}
\alias{plot_gather_classified}
\title{Visualize the classified and unclassified fraction of many samples using sourmash gather results.}
\usage{
plot_gather_classified(gather_df)
}
\arguments{
\item{gather_df}{A data frame of multiple sourmash gather results created by `read_gather()`.}
}
\value{
A ggplot2 plot.
}
\description{
`plot_gather_classified()` produces a ggplot2 plot of the classified and unclassified fraction of many samples from a data frame of sourmash gather results.
}
\examples{
\dontrun{
plot_gather_classified(gather_df)
}
}
