% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signatures.R
\name{plot_signatures_rarefaction}
\alias{plot_signatures_rarefaction}
\title{Plot a rarefaction curve to determine sequence saturation}
\usage{
plot_signatures_rarefaction(rarefaction_df, fraction_of_points_to_plot = 500)
}
\arguments{
\item{rarefaction_df}{A tidy data frame returned by `from_signatures_to_rarefaction_df()`. Must contain the columns `num_kmers_sampled`, `num_kmers_observed`.}

\item{fraction_of_points_to_plot}{Integer. 1 every `fraction_of_points_to_plot` will be plotted.}
}
\value{
A ggplot2 object.
}
\description{
`plot_signatures_rarefaction()` produces a ggplot2 plot from the tidy data frame output by `from_signatures_to_rarefaction_df()`.
Since the rarefaction curves may contain many dense points, the `fraction_of_points_to_plot` argument reduces the total number of points rendered which decreases the functions run time.
Rarefaction curves that end with a shallow slow likely represent samples that were sequenced deeply enough to capture the k-mer diversity in the sequence.
}
\examples{
plot_signatures_rarefaction(rarefaction_df)
}
