% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signatures.R
\name{plot_signatures_upset}
\alias{plot_signatures_upset}
\title{Visualize the intersection of minhashes in sourmash signatures from many samples}
\usage{
plot_signatures_upset(upset_df, ...)
}
\arguments{
\item{upset_df}{An upset-compliant data frame.}

\item{...}{Arguments passed to ComplexUpset::upset().}
}
\value{
An upset plot produced by ComplexUpset::upset().
}
\description{
`plot_signatures_upset` uses `ComplexUpset::upset()` to plot the intersection of minhashes in sourmash signatures from many samples.
While based on ggplot2, `ComplexUpset::upset()` has a difficult syntax to parameterize.
To produce an alternative visualization, it may be easiest to run `sourmashconsumr::plot_signatures_upset` to retrieve the function source code and alter it.
}
\examples{
\dontrun{
plot_signatures_upset()
}
}
