% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxonomy_annotate.R
\name{plot_taxonomy_annotate_sankey}
\alias{plot_taxonomy_annotate_sankey}
\title{Visualize a sankey diagram from taxonomic lineages from one or many samples}
\usage{
plot_taxonomy_annotate_sankey(
  taxonomy_annotate_df,
  tax_glom_level = NULL,
  palette = NULL,
  label = TRUE
)
}
\arguments{
\item{taxonomy_annotate_df}{Data frame containing outputs from sourmash taxonomy annotate. Can contain results from one or many runs of sourmash taxonomy annotate. If specified, agglomeration occurs across all queries.}

\item{tax_glom_level}{Optional character string specifying the taxonomic rank to agglomerate k-mer counts. Must be one of "domain", "phylum", "class", "order", "family", "genus", "species."}

\item{palette}{Optional character vector specifying a palette. Colors in the palette are recycled across taxonomic labels. If no palette is specified, RColorBrewer's Set2 is the default.}

\item{label}{Boolean controlling whether taxonomy labels are added to the plot. The default, TRUE, plots labels. Setting to FALSE removes the labels and can be used to control the label appearances; add a layer with `ggforce::geom_parallel_sets_labels()` to re-add labels and use arguments passed to `ggplot2::layer()` to control the appearance of the output.}
}
\value{
A ggplot2 plot
}
\description{
`plot_taxonomy_annotate_sankey()` plots a sankey diagram from the output of sourmash taxonomy annotate.
The input data frame can contain one or many samples.
If there are many samples, abundances of each lineage are summarized and sample-level information is lost.
If the parameter `tax_glom_level` is specified, the plot will be summarized to that taxonomic rank (e.g. if "order" is specified, only domain, phylum, class, and order will be plotted).
}
\examples{
\dontrun{
plot_taxonomy_annotate_sankey()
}
}
