% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxonomy_annotate.R
\name{plot_taxonomy_annotate_upset}
\alias{plot_taxonomy_annotate_upset}
\title{Visualize an upset plot of taxonomic lineage intersections between samples}
\usage{
plot_taxonomy_annotate_upset(upset_inputs, fill = NULL)
}
\arguments{
\item{upset_inputs}{List of inputs produced by from_taxonomy_annotate_to_upset_inputs().}

\item{fill}{Optional argument specifying which level of taxonomy to fill the upset plot intersections with. Uses the Set2 palette so cannot visualize more than 8 levels.}
}
\value{
A ComplexUpset plot
}
\description{
`plot_taxonomy_annotate_upset()` produces a ComplexUpset plot displaying the intersection of taxonomic lineages observed in many samples.
The bar chart that displays the intersection size between samples can optionally be colored by taxonomy lineage (e.g. phylum).
}
\examples{
\dontrun{
plot_taxonomy_annotate_upset()
}
}
