% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sourmash_outputs.R
\name{read_gather}
\alias{read_gather}
\title{Read CSV file or files output by sourmash gather}
\usage{
read_gather(file, intersect_bp_threshold, ...)
}
\arguments{
\item{file}{Path to CSV file or files output by sourmash gather.}

\item{intersect_bp_threshold}{Integer. Gather matches must have an intersect_bp greater than or equal to this value.}

\item{...}{Arguments passed to read_csv().}
}
\value{
A tibble.
}
\description{
Read CSV file or files output by sourmash gather
}
\examples{
read_gather()
}
