% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sourmash_outputs.R
\name{read_signature}
\alias{read_signature}
\title{Read sourmash signatures into a dataframe}
\usage{
read_signature(file, compliant = TRUE)
}
\arguments{
\item{file}{Path to signature (json) file output by sourmash sketch (previously sourmash compute).}

\item{compliant}{Boolean indicating whether signature columns should be compliant; the json fields changed across versions of sourmash. This may drop deprecated columns like 'type' but will allow you to bind many signatures into a single data frame even if they were sketched with different versions of sourmash.}
}
\value{
A tibble.
}
\description{
Read sourmash signatures into a dataframe
}
\examples{
\dontrun{
read_signature("tests/testthat/SRR18071810.sig")
}
}
