% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sourmash_outputs.R
\name{read_signature_one}
\alias{read_signature_one}
\title{Read one sourmash signature into a dataframe}
\usage{
read_signature_one(file, compliant = TRUE)
}
\arguments{
\item{file}{Path to signature (json) file output by sourmash sketch (previously sourmash compute).}

\item{compliant}{Boolean indicating whether signature columns should be compliant; the json fields changed across versions of sourmash. This may drop deprecated columns like 'type' but will allow you to bind many signatures into a single data frame even if they were sketched with different versions of sourmash.}
}
\value{
A tibble.
}
\description{
`read_signature_one` reads one signature into a data frame.
It's mostly a dummy function, but breaking it out this way allows one or many files to be passed to `read_signature`.
}
\examples{
\dontrun{
read_signature("tests/testthat/SRR18071810.sig")
}
}
