% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder.R
\name{pivot_wider_taxonomy_annotate}
\alias{pivot_wider_taxonomy_annotate}
\title{Pivot sourmash_taxonomy_results to wide format}
\usage{
pivot_wider_taxonomy_annotate(taxonomy_annotate_df)
}
\arguments{
\item{taxonomy_annotate_df}{}
}
\value{
A data frame in wide format. Variables expanded is n_unique_kmers and colnames are query_name.
}
\description{
Pivot sourmash_taxonomy_results to wide format
}
\examples{
pivot_wider_taxonomy_annotate(taxonomy_annotate_df)
}
