% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sourmash_outputs.R
\name{read_compare_csv}
\alias{read_compare_csv}
\title{Read a CSV file output by sourmash compare}
\usage{
read_compare_csv(file, sample_to_rownames = F, ...)
}
\arguments{
\item{file}{path to file output by sourmash compare using the with the --csv flag}

\item{sample_to_rownames}{Boolean indicating whether sample names should be added to the tibble as a column or a rowname}

\item{...}{Arguments passed to readr::read_csv()}
}
\value{
A tibble.
}
\description{
Read a CSV file output by sourmash compare
}
\examples{
\dontrun{
read_compare_csv("")
}
}
