% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sourmash_outputs.R
\name{read_gather}
\alias{read_gather}
\title{Read CSV file or files output by sourmash gather}
\usage{
read_gather(file, intersect_bp_threshold = 0, ...)
}
\arguments{
\item{file}{Path to CSV file or files output by sourmash gather.}

\item{intersect_bp_threshold}{Integer. Gather matches must have an intersect_bp greater than or equal to this value.}

\item{...}{Arguments passed to read_csv().}
}
\value{
A tibble.
}
\description{
`read_gather()` reads in one or many CSV files produced by the command line function sourmash taxonomy annotate.
Genome matches can be filtered with `intersect_bp_threshold`, whose default value is 0 base pairs.
}
\examples{
\dontrun{
read_gather()
}
}
