% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxonomy_annotate.R
\name{tax_glom_taxonomy_annotate}
\alias{tax_glom_taxonomy_annotate}
\title{Agglomerate counts of same lineage to specified level of taxonomy.}
\usage{
tax_glom_taxonomy_annotate(taxonomy_annotate_df, tax_glom_level = NULL)
}
\arguments{
\item{taxonomy_annotate_df}{Data frame containing outputs from sourmash taxonomy annotate. Can contain results from one or many runs of sourmash taxonomy annotate. Agglomeration occurs within each query.}

\item{tax_glom_level}{Character. NULL by default, meaning no agglomeration is done. Valid options are "domain", "phylum", "class", "order", "family", "genus", and "species". When a valid option is supplied, k-mer counts are agglomerated to that level}
}
\value{
A data frame.
}
\description{
Inspired by phyloseq::tax_glom(), this method summarizes k-mer counts from genomes that have the same taxonomy at a user-specified taxonomy rank.
Agglomeration occurs within each sample, meaning the number of k-mers is only summed within each query_name.
This function returns a data frame with the columns `lineage`, `query_name`, and `n_unique_kmers`.
}
\examples{
\dontrun{
tax_glom_taxonomy_annotate()
}
}
