% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder.R
\name{taxonomy_annotate_to_metacoder}
\alias{taxonomy_annotate_to_metacoder}
\title{taxonomy_annotate_to_metacoder}
\usage{
taxonomy_annotate_to_metacoder(
  taxonomy_annotate_tibble = NULL,
  file = NULL,
  intersect_bp_threshold = 50000,
  summary_level = c(NULL, "domain", "phylum", "class", "order", "family", "genus",
    "species"),
  groups = NULL,
  groups_prefix = "x",
  ...
)
}
\arguments{
\item{taxonomy_annotate_tibble}{Tibble containing outputs from sourmash taxonomy annotate. If specified, file is ignored. Can contain results from one or many runs of sourmash taxonomy annotate.}

\item{file}{Path to CSV file or files output by sourmash taxonomy annotate.}

\item{intersect_bp_threshold}{Integer. Gather matches must have an intersect_bp greater than or equal to this value.}

\item{summary_level}{Character.}

\item{groups}{A tibble or data.frame with distinct query_name values from taxonomy_annotate_tibble in the first column and query groups in the second column.}

\item{groups_prefix}{Character. Ignored if groups is defined. Used to prefix query_name values for the metacoder function calc_n_samples().}

\item{...}{Arguments passed to metacoder::parse_tax_data().}
}
\value{
A metacoder taxmap object.
}
\description{
taxonomy_annotate_to_metacoder
}
\examples{
taxonomy_annotate_to_metacoder()
}
